bitkeeper revision 1.507 (3f8ad9d51NEWhiPGDd2jxu-ez0fnOA)
i386_ksyms.c, sched.h, ide-cd.h, memory.c, traps.c, process.c, entry.S:
  Fix LDT bug when switching domains.

+++ xen/drivers/ide/ide-cd.h
@@ -439,7 +439,7 @@ struct atapi_mechstat_header {
 
 	byte     curlba[3];
 	byte     nslots;
-	__u8 short slot_tablelen;
+	__u16 slot_tablelen;
 };
